Wwise Sidechain Demo Template Using Wwise Meter and EQ
This is a template using wwise meter to sidechain duck music and sound
effects while dialogue occurs. This is a smoother type of ducking.
I want to share to encourage others to share their own templates,
solutions, and Wwise tricks with the community so we stop solving the
same problems over and over. This will free us up to focus more on our
art and not the barriers in the way :)